requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
drf-spectacular>=0.27
pymongo>=4.6
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.2
uvicorn>=0.23
